Fix writing to the floppy image.
The previous code only worked by accidental timing.
It started a gcr-bit-counter (gcr_bit_cnt) when the DOS turns on the
write head. Then every 10 bit times (corresponding to 8 binary bits) it
would check if a '07' byte was written (decoded from the GCR which is
actually passed from the VIA's parallel port). That is the value that
comes after a sync which indicates it is the data part of the block, as
opposed to the header.

This didn't work with the 2031 DOS ROM. It seemed it was approximately 2
bit times slower in starting to write the sync after turning on the
write head. As a result, the decoding was out of sync and the '07' was
never detected.

I added an extra counter that counts sync bytes as they are written.
Better would be to count bits, so this is also a shortcut that happens
to work with DOS (which writes SYNC marks in multiples of 8 bits).
If at least 2 bytes (i.e. 16 bits) of SYNC are detected, and then
something else is written, restart the bit counting (and some other
counters, which may or may not be strictly needed).

Now the '07' is detected as it is written.

(Also I retabbed the file; tabstops of 3 are very unusual)