Wire in the CRTC and allow switching with discrete video.
Use the sync signals also as the blank signals.
The discrete video still works.
The CRTC produces an image but it is incomplete. The top/bottom borders
seem too large. The letters are very wide and the first 6 of each line
are missing. (This width is about the same as the extra space on the
wide-screen monitor; coincidence?)