Pass 2 of giving each drive a separate track buffer.
Only read track when the track# changes

It even kind of works.

I may of course have broken other things I did not notice yet.

*** With this commit a drive-to-drive file copy results in lit error-led
during the copy, several times; mostly retries by DOS seem to succeed in
the end, but eventually it fails with an error anyway: 25 WRITE ERROR
which is a data-verify error.