Fix 8250 read-after-write 21,read error case.
`ieeedrv_track` outputs `busy` while it is reading into or writing from
its track buffer (from/to the QNice disk image buffer).

`ieeedrv_trkgen` inputs `busy` and as long as it is true, does not
advance the `bit_clk_cnt` which in essence means that the entire state
machine stops.

`ieeedrv_track` wants to save its track buffer when the FDC has written
some data into it, so it initiates a write in that case.

When the 8250 FDC rom as written a sector, it wants to read it back to
check it was written correctly.

By bad luck, at the same time `ieeedrv_track` has set `busy` since it is
writing out its track buffer to QNice.

With the state machine in `ieeedrv_trkgen` stopped, the FDC doesn't even
receive SYNC signals, and times out, resulting in FDC error #3 which
corresponds to DOS error 21.

Solution: stop the clock for the FDC CPU and VIA, so that they do not
notice that time is passing while there is track buffer activity going
on behind their back.