Almost eliminate the drv_sel -> drv_act delay
The drive controller's IRQ switches active drives for a brief moment to
check for the write protect switch. This messed up the track data
buffering though, which could not switch track contents that quickly
even in the MiSTeR. So the delay was introduced.

With 2 track buffers, there should be no need any more for this.

Now on the 8250 we see a lot of left-to-right flashing on first
disk access on a D80 image. Maybe the combination of the IRQ drive
switch, the check of the second side, and the other activities that are
mixed into the LED.

*** Still doesn't fix the drive-to-drive copying problems.