Use up-to-date selected drive when it changes.
The selected drive has changed, so we can't use drv_act,
but must use drv_sel_s. Likewise, we can not use
ltrack_new which is also based on drv_act.
So we open-code the value of ltrack_new, but using the up-to-date
drv_sel_s.
This is needed twice in ieeedrv_track.