Add a trivial SuperPET board.
It plugs into the CPU socket. So far this board only passes on the
6502's signals.

Since it plugs only into the CPU socket, the main board is taken totally
unaware. No changes should be necessary. All extra SuperPET hardware is
present on the board and will be seen by either CPU. In that case a
dummy address is placed on the address bus as fed to the main board.

We make one compromise: we want to use the already existing 64 KB of
expansion memory also for the SuperPET. To that end, we add an extra
address line (addr[16]) which serves as a selector for this extension
ram.